+++ src/cart.html
@@ -82,5 +82,6 @@
     </div>
 </body>
 
-<script src='./js/product.js'></script>
+<script src='./js/Product.js'></script>
+<script>showResult2()</script>
 </html>
