+++ src/product.html
@@ -93,6 +93,7 @@
 
 
 <script src='./js/main.js'></script>
-<script src='./js/product.js'></script>
+<script src='./js/Product.js'></script>
+<script>showResult()</script>
 
 </html>
